% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmvlr}
\alias{cmvlr}
\title{Fit the multivariate logistic regression model with Gibbs sampling}
\usage{
cmvlr(x, y, iter, beta0, r0, phi0, mubeta, sigbeta, omega, burn_in,
  thinning, burn_in_truncated = 10L)
}
\arguments{
\item{x}{numeric array containing the covariates for each sample}

\item{y}{numeric matrix containing the outcomes of interest}

\item{iter}{integer, the number of iterartions}

\item{beta0}{intial values for the coefficients}

\item{r0}{initial values for the unique entries of the scale matrix}

\item{phi0}{initial values for phis}

\item{mubeta}{prior mean on beta}

\item{sigbeta}{prior covariance matrix for beta}

\item{omega}{variance-covariance matrix for sampling r (Metropolis step)}

\item{burn_in}{number of burn in samples to use for complete loop}

\item{thinning}{thinning used}

\item{burn_in_truncated}{number of burn in samples to use for sampling from
the truncated normal distribution}
}
\description{
Fit the multivariate logistic regression model with Gibbs sampling
}
