% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvlogis}
\alias{dmvlogis}
\title{Calculate the multivariate logistic density of a vector}
\usage{
dmvlogis(x, mean, R, df = 7.3, logd = FALSE)
}
\arguments{
\item{x}{numeric vector containing the random numbers}

\item{mean}{numeric vector}

\item{R}{numeric scale matrix with 1's on the diagonal}

\item{df}{degrees of freedom}

\item{logd}{boolean indicating whether the log-transformed density
should be returned}
}
\description{
Calculate the multivariate logistic density of a vector
}
