% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvt}
\alias{dmvt}
\title{Calculate the multivariate Student t density of a vector}
\usage{
dmvt(x, mean, sigma, df, logd = FALSE)
}
\arguments{
\item{x}{numeric vector containing the random numbers}

\item{mean}{numeric vector}

\item{sigma}{numeric covariance matrix}

\item{df}{degrees of freedom}

\item{logd}{boolean indicating whether the log-transformed density
should be returned}
}
\description{
Calculate the multivariate Student t density of a vector
}
