\name{multlogreg-package}
\alias{multlogreg-package}
\alias{multlogreg}
\docType{package}
\title{
\packageTitle{multlogreg}
}
\description{
\packageDescription{multlogreg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multlogreg}
\packageIndices{multlogreg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{multlogreg}

Maintainer: \packageMaintainer{multlogreg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
