% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multlogreg-package.R
\docType{package}
\name{multlogreg}
\alias{multlogreg}
\alias{multlogreg-package}
\title{multlogreg: Multivariate Logistic Regression Models}
\description{
Fit multivariate logistic regression models.
}
