% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlr.R
\name{mvlr}
\alias{mvlr}
\title{Fit the multivariate logistic regression model with Gibbs sampling}
\usage{
mvlr(coeff, data, iter, beta0, r0, phi0, mubeta, sigbeta, omega, burn_in,
  thinning, burn_in_truncated = 10)
}
\arguments{
\item{coeff}{named list of coefficients for each outcome,
intercepts are included by default}

\item{data}{data.frame containing the outcomes and coefficients}

\item{iter}{integer, the number of iterartions}

\item{mubeta}{prior mean on beta}

\item{sigbeta}{prior covariance matrix for beta}

\item{omega}{variance-covariance matrix for sampling r (Metropolis step)}

\item{burn_in}{number of burn in samples to use for complete loop}

\item{thinning}{thinning used}

\item{burn_in_truncated}{number of burn in samples to use for sampling from
the truncated normal distribution}

\item{betainit}{intial values for the coefficients}

\item{rinit}{initial values for the unique entries of the scale matrix}

\item{phiinit}{initial values for phis}
}
\value{
A list with the samples of the coefficients,
  correlation coefficients, weights and the number of accepted values
}
\description{
Fit the multivariate logistic regression model with Gibbs sampling
}
