% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtmvn_accrej}
\alias{rtmvn_accrej}
\title{Generate samples from a truncated multivariate normal distribution}
\usage{
rtmvn_accrej(n, mean, sigma, D, lower, upper, nproposals = 0L,
  verbose = FALSE)
}
\arguments{
\item{n}{number of samples to be generated}

\item{mean}{mean vector (q x 1)}

\item{sigma}{covariance matrix (q x q)}

\item{D}{matrix of linear constraints}

\item{lower}{vector of lower bounds}

\item{upper}{vector of upper bounds}

\item{nproposals}{number of propsal samples per iteration}

\item{verbose}{should information about sampling be printed?}
}
\description{
Generates a n x q matrix with n samples of a truncated multivariate
normal distribution with dimension q using the acceptance-rejection method
sampling from a multivariate normal distribution.
}
\details{
For the a non-diagonal matrix D, the linear restriction lower <= Dx <= upper
hold.
}
