% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtmvn_gibbs}
\alias{rtmvn_gibbs}
\title{Gibbs sampler for the Truncated Multivariate Normal Distribution}
\usage{
rtmvn_gibbs(n, mean, sigma, D, lower, upper, init, burn_in = 10L,
  thinning = 1L)
}
\arguments{
\item{n}{number of samples to be generated}

\item{mean}{mean vector}

\item{sigma}{covariance matrix}

\item{D}{matrix of linear constraints}

\item{lower}{vector of lower bounds}

\item{upper}{vector of upper bounds}

\item{init}{vector of initial values for the Gibbs sampler.}

\item{burn_in}{number of burn in samples to use.}

\item{thinning}{thinning parameter}
}
\value{
matrix of samples with each column being an idependent sample.
}
\description{
Random vector generation for the truncated multivariate normal distribution
    using a Gibbs sampler.
}
\references{
Li, Y., & Ghosh, S. K. (2015). Efficient sampling methods for
    truncated multivariate normal and student-t distributions subject to
    linear inequality constraints. Journal of Statistical Theory and
    Practice, 9(4), 712-732.
}
