% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtuvn_single}
\alias{rtuvn_single}
\title{Univariate Truncated Normal Distribution - single sample}
\usage{
rtuvn_single(mean, sd, lower, upper)
}
\description{
Generate one sample from the truncated normal distribution
 TN(mean, sd; lower, upper).
}
